+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - community: clojure pack & sub-menu descriptions
 - practicalli: personal preferences with conditional skip
 - practicalli: startup dashboard logo
+- community: proffered plugins from astrocommunity
 
 ## Changed
 - lazy: enable zipPlugin to view src inside jar files